Make code pretty
