+++ src/components/Navbar/index.ts
@@ -1,0 +1,2 @@
+import Navbar from "./Navbar";
+export default Navbar;
